Module E.SetEncoding

val name : string

Name of the encoding used for debugging.

val rewrite : SMT.t -> SMT.t

Rewrite all set symbols and operations to the encoding.

val rewrite_back : SMT.t -> SMT.Model.t -> SMT.Model.t

Rewrite a model back from the encoding. The original formula is provided to reconstruct sorts of variables.