Module type Logic_sig.LOGIC

module Sort : SORT

module Variable : VARIABLE with module Sort = Sort

type t

type term

include Datatype_sig.PRINTABLE with type t := t

include Datatype_sig.SHOW with type t := t

val pp : Stdlib.Format.formatter -> t -> unit

Output to formatter

val print : ?prefix:string -> t -> unit

Output to stdout.

val show_option : t option -> string

val print_option : ?prefix:string -> t option -> unit

val dump : string -> t -> unit

Dump to file given by filename

Lists of printable values

val show_list : ?separator:string -> t list -> string

val pp_list : Stdlib.Format.formatter -> t list -> unit

val print_list : ?separator:string -> ?prefix:string -> t list -> unit

include Datatype_sig.COMPARABLE with type t := t

include Datatype_sig.COMPARISON with type t := t

include Datatype_sig.COLLECTIONS with type t := t

module Set : Datatype_sig.SET with type elt = t

Set over type t

module MonoList : Datatype_sig.MONO_LIST with type key = t and type t = t list

Monomorphic list over type t

module Map : Datatype_sig.MAP with type key = t

Polymorphic map from t to 'a

module MonoMap (Data : Datatype_sig.SHOW) : Datatype_sig.MONO_MAP with type key = t and type data = Data.t

Monomorphic map from t to Data.t

include SORTED with type t := t and module Sort := Sort

val get_sort : t -> Sort.t

val has_sort : Sort.t -> t -> bool

val show_with_sort : t -> string

val hash : t -> int

val mk_var : string -> Sort.t -> t

val mk_fresh_var : string -> Sort.t -> t

val get_all_sorts : t -> Sort.t list

val of_var : Variable.t -> t

val of_const : Constant.t -> t

val get_sort_in_term : string -> t -> Sort.t

val get_operands : t -> t list

val is_quantifier_free : t -> bool

val is_ground : ground:Variable.t list -> t -> bool

val is_ground' : forbidden:Variable.t list -> t -> bool

True if the term does not contatin any of provided variables.

val show : t -> string

val equal : t -> t -> bool

val compare : t -> t -> int

val map : (t -> t) -> t -> t

val map_terms : (term -> term) -> t -> t

val get_vars : t -> Variable.t list

Return all variables, without duplicates.

val free_vars : t -> Variable.t list

val bound_vars : t -> Variable.t list

val rename_var : string -> string -> t -> t

val exists : (t -> bool) -> t -> bool

val for_all : (t -> bool) -> t -> bool

val select_subformulae : (t -> bool) -> t -> t list

val substitute : t -> var:Variable.t -> by:term -> t

val substitute_list : t -> vars:Variable.t list -> by:term list -> t

val replace_subformula : t -> subformula:t -> by:t -> t

val size : t -> int

type ast

val to_ast : ?dagify:bool -> t -> ast

val output_ast : string -> ast -> unit

val to_smt2 : t -> string

Human readable SMT-LIB string.

val to_sexp : t -> Sexplib.Sexp.t

S-expression representation used to generate SMT-LIB string. Usefull when printing a formula embeded in a larger s-expr.

type pred_sigs := (string * Sort.t list) list

val to_bench : ?pred_sigs:pred_sigs -> ?source:string -> ?status:[< `Sat | `Unsat | `Unknown ] -> ?options:string -> t -> string

val output_benchmark : ?pred_sigs:pred_sigs -> ?source:string -> ?status:[< `Sat | `Unsat | `Unknown ] -> ?options:string -> string -> t -> unit

val (===) : t -> t -> bool