Astral

Astral is a SMT solver for separation logics.

Solver

Astral.Solver

Pure formulae

Astral.Constant

Operations over constant values.

Astral.SMT

Constructors and operations over pure formulae.

Separation Logic

Astral.Sort

Astral.MemoryModel

Access to Astral's memory model.

Astral.SL

Basic operations over separation logic formulae.

Astral.SLID

Operations over separation logic which take into account semantics of inductive predicates.

Astral.SL_builtins

Constructors for built-in definitions provided by Astral.

Astral.StackHeapModel

Stack-heap models of separation logic.

Frontends

Astral.LowLevelSeplog

Low-level separation logic with block operations, interpreted over bitvectors.